refactor: 대기 영문 오타 수정 (wating -> waiting)
