+++ src/main/java/roomescape/waiting/domain/exception/PastDateTimeWaitingException.java
@@ -1,0 +1,10 @@
+package roomescape.waiting.domain.exception;
+
+import roomescape.common.exception.UnprocessableContentException;
+
+public class PastDateTimeWaitingException extends UnprocessableContentException {
+
+    public PastDateTimeWaitingException() {
+        super("과거 시간의 예약에 대기를 등록할 수 없습니다.");
+    }
+}
