+++ src/main/java/roomescape/waiting/domain/exception/WaitingNotFoundException.java
@@ -1,0 +1,10 @@
+package roomescape.waiting.domain.exception;
+
+import roomescape.common.exception.NotFoundException;
+
+public class WaitingNotFoundException extends NotFoundException {
+
+    public WaitingNotFoundException() {
+        super("존재하지 않는 대기입니다.");
+    }
+}
