+++ src/main/java/roomescape/waiting/domain/exception/WaitingSlotDuplicateException.java
@@ -1,0 +1,10 @@
+package roomescape.waiting.domain.exception;
+
+import roomescape.common.exception.ConflictException;
+
+public class WaitingSlotDuplicateException extends ConflictException {
+
+    public WaitingSlotDuplicateException() {
+        super("해당 시간에 이미 대기가 존재합니다.");
+    }
+}
