+++ src/main/java/roomescape/waiting/service/dto/response/WaitingCreateResponse.java
@@ -1,0 +1,6 @@
+package roomescape.waiting.service.dto.response;
+
+public record WaitingCreateResponse(
+    Long id
+) {
+}
