--- src/main/java/roomescape/wating/service/dto/response/WaitingCreateResponse.java
@@ -1,6 +1,0 @@
-package roomescape.wating.service.dto.response;
-
-public record WaitingCreateResponse(
-    Long id
-) {
-}
